test(filters): consolidate search regressions with existing tests
Move the search SQL regression matrix into the existing filter test module and reuse its movie model factory instead of maintaining a standalone unit-test module and duplicate model. Keep compilation checks independent of database fixtures so Spanner coverage still runs without the skipped integration backend.

All 154 selected filter tests and the full 292-file mypy check pass. Runtime behavior is unchanged.
