+++ tests/integration/test_filters.py
@@ -3,6 +3,7 @@ from datetime import datetime, timezone
 from typing import Any, Literal, Optional
 
 import pytest
+from google.cloud.sqlalchemy_spanner.sqlalchemy_spanner import SpannerDialect
 from pytest import FixtureRequest
 from sqlalchemy import Boolean, Engine, ForeignKey, String, create_engine, func, select
 from sqlalchemy.dialects import mssql, mysql, oracle, postgresql, sqlite
@@ -807,6 +808,42 @@ def test_search_filter(session: Session, movie_model_sync: type[DeclarativeBase]
     statement = search_filter.append_to_statement(select(Movie), Movie)
     results = session.execute(statement).scalars().all()
     assert len(results) == 1
+
+
+@pytest.mark.unit
+@pytest.mark.parametrize(
+    "dialect",
+    [
+        sqlite.dialect(),  # type: ignore[no-untyped-call]
+        postgresql.dialect(),  # type: ignore[no-untyped-call]
+        mysql.dialect(),  # type: ignore[no-untyped-call]
+        oracle.dialect(),  # type: ignore[no-untyped-call]
+        mssql.dialect(),  # type: ignore[no-untyped-call]
+        SpannerDialect(),
+    ],
+)
+@pytest.mark.parametrize(
+    ("filter_type", "ignore_case", "operator"),
+    [
+        (SearchFilter, False, "like"),
+        (SearchFilter, True, "ilike"),
+        (NotInSearchFilter, False, "not_like"),
+        (NotInSearchFilter, True, "not_ilike"),
+    ],
+)
+@pytest.mark.parametrize("value", ["plain", "50%_/"])
+def test_search_filter_preserves_default_sql(
+    dialect: Any, filter_type: type[SearchFilter], ignore_case: bool, operator: str, value: str
+) -> None:
+    Movie = get_movie_model_for_engine("sqlite", "search_compile")
+    expected = select(Movie).where(getattr(Movie.title, operator)(f"%{value}%")).compile(dialect=dialect)
+    for search_filter in (
+        filter_type("title", value, ignore_case=ignore_case),
+        filter_type("title", value, ignore_case=ignore_case, escape_wildcards=False),
+    ):
+        compiled = search_filter.append_to_statement(select(Movie), Movie).compile(dialect=dialect)
+        assert str(compiled) == str(expected)
+        assert compiled.params == expected.params
 
 
 def test_search_filter_escapes_wildcards(session: Session, movie_model_sync: type[DeclarativeBase]) -> None:
